Renaming Basiccoin => Tokendeploy
