+++ js/src/dapps/tokendeploy/IdentityIcon/identityIcon.js
