+++ js/src/dapps/tokendeploy/IdentityIcon/index.js
