+++ js/src/dapps/tokendeploy/parity.js
