+++ js/src/views/Dapps/builtin.json
@@ -1,7 +1,7 @@
 [
   {
     "id": "0xf9f2d620c2e08f83e45555247146c62185e4ab7cf82a4b9002a265a0d020348f",
-    "url": "basiccoin",
+    "url": "tokendeploy",
     "name": "Token Deployment",
     "description": "Deploy new basic tokens that you are able to send around",
     "author": "Parity Team <admin@ethcore.io>",
